requirements.txt
Flask
flask-cors
dnspython
elasticsearch>=6.0.0,<7.0.0
gunicorn
